backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
nixtla
statsforecast
python-multipart
python-dotenv
pytest
httpx
